+++ e2e/nx-firebase-e2e/test-utils/utils.ts
@@ -1,0 +1,420 @@
+
+import { joinPathFragments, names } from '@nx/devkit'
+import { readJson, runNxCommandAsync } from '@nx/plugin/testing'
+import { expectStrings, expectNoStrings } from './test-helpers'
+import { testDebug, red, green } from './test-logger'
+
+const NPM_SCOPE = '@proj'
+const STRIP_ANSI_MATCHER = /[\u001b\u009b][[()#;?]*(?:[0-9]{1,4}(?:;[0-9]{0,4})*)?[0-9A-ORZcf-nqry=><]/g
+
+export interface ProjectData {
+  name: string
+  directory: string // from Nx 16.8.0 this is the apps/libs prefix
+  projectName: string
+  projectDir: string
+  srcDir: string
+  distDir: string
+  mainTsPath: string
+  npmScope: string
+  configName: string
+}
+
+
+export async function safeRunNxCommandAsync(cmd: string)
+{
+  testDebug(`- safeRunNxCommandAsync ${cmd}`)
+  try {
+    async function runCommand(cmd: string) {
+      const result = await runNxCommandAsync(`${cmd} --verbose`, { silenceError: true })
+      // strip chalk TTY ANSI codes from output
+      result.stdout = result.stdout.replace(STRIP_ANSI_MATCHER, '')
+      result.stderr = result.stderr.replace(STRIP_ANSI_MATCHER, '')   
+      if (result.stdout) {
+        testDebug(green(result.stdout))
+      }
+      if (result.stderr) {
+        testDebug(red(result.stderr))
+      }
+      return result
+    }
+    // getting wierd lock file errors from Nx, so retry at least once
+    let result = await runCommand(cmd)
+    if (result.stdout.includes('LOCK-FILES-CHANGED') || result.stderr.includes('LOCK-FILES-CHANGED')) {
+      testDebug(red(`Re-running command ${cmd} due to LOCK-FILES-CHANGED`))
+      result = await runCommand(cmd)
+    }
+
+    return result
+  }
+  catch (e) {
+    testDebug(red(`ERROR: Running command ${(e as Error).message}`))
+    throw e 
+  }
+}
+
+export async function runTargetAsync(projectData: ProjectData, target: string = 'build') {
+
+  //SM: Mar'24 - this seems legacy, not sure if needed
+  // if (target === 'build') {
+  //     // need to reset Nx here for e2e test to work
+  //     // otherwise it bundles node modules in the main.js output too
+  //     // I think this is a problem with dep-graph, since it works if main.ts
+  //     // is modified before first build      
+  //     await runNxCommandAsync('reset')    
+  // }
+  if (target === 'build') {
+    // getting wierd errors with Nx 16.8.1 where is says it cannot find the project
+    // need to reset Nx here for e2e test to work
+    // I dont think the Nx daemon has enough time to update its cache
+    // after generation of a new project and building it right away
+    await runNxCommandAsync('reset')    
+  }  
+
+  testDebug(`- runTargetAsync ${target} ${projectData.projectName}`)
+  const result = await safeRunNxCommandAsync(`${target} ${projectData.projectName}`)
+  // testDebug(result.stdout)
+  // testDebug(result.stderr)
+
+  if (target === 'build') {
+    expectStrings(result.stdout, [
+      `Successfully ran target ${target} for project ${projectData.projectName}`
+    ])   
+  }
+
+  return result 
+}
+
+export async function removeProjectAsync(projectData: ProjectData) {
+  const result = await safeRunNxCommandAsync(`g @nx/workspace:remove ${projectData.projectName} --forceRemove`)
+  expectStrings(result.stdout, [
+    `DELETE ${projectData.projectDir}/project.json`,
+    `DELETE ${projectData.projectDir}`,
+  ])   
+  return result 
+}
+
+export async function renameProjectAsync(projectData: ProjectData, renameProjectData: ProjectData) {
+  //TODO: this wont work if destination project is in a subdir
+  const result = await safeRunNxCommandAsync(`g @nx/workspace:move --project=${projectData.projectName} --destination=${renameProjectData.projectName}`)
+  expectStrings(result.stdout, [
+    `DELETE apps/${projectData.projectName}/project.json`,
+    `DELETE apps/${projectData.projectName}`,
+    `CREATE apps/${renameProjectData.projectName}/project.json`,
+  ])   
+  return result 
+}
+
+export async function appGeneratorAsync(projectData: ProjectData, params: string = '') {
+  testDebug(`- appGeneratorAsync ${projectData.projectName} ${params}`)
+  const result = await safeRunNxCommandAsync(`g @simondotm/nx-firebase:app ${projectData.name} --directory=${projectData.directory} ${params}`)
+  // testDebug(result.stdout)
+
+  return result
+}
+
+export async function functionGeneratorAsync(projectData: ProjectData, params: string = '') {
+  testDebug(`- functionGeneratorAsync ${projectData.projectName} ${params}`)
+  const result = await safeRunNxCommandAsync(`g @simondotm/nx-firebase:function ${projectData.name} --directory=${projectData.directory} ${params}`)
+  // testDebug(result.stdout)
+  // testDebug(result.stderr)
+  return result
+}
+
+export async function libGeneratorAsync(projectData: ProjectData, params: string = '') {
+  testDebug(`- libGeneratorAsync ${projectData.projectName}`)
+  const result = await safeRunNxCommandAsync(`g @nx/js:lib ${projectData.name} --directory=${projectData.directory} --projectNameAndRootFormat=derived ${params}`)
+  // testDebug(result.stdout)
+  // testDebug(result.stderr)  
+  return result
+}
+
+export async function syncGeneratorAsync(params: string = '') {
+  testDebug(`- syncGeneratorAsync ${params}`)
+  return await safeRunNxCommandAsync(`g @simondotm/nx-firebase:sync ${params}`)
+}
+
+export async function migrateGeneratorAsync(params: string = '') {
+  testDebug(`- migrateGeneratorAsync ${params}`)
+  return await safeRunNxCommandAsync(`g @simondotm/nx-firebase:migrate ${params}`)
+}
+
+
+
+export async function cleanAppAsync(projectData: ProjectData, options = { appsRemaining:0, functionsRemaining: 0}) {
+  testDebug(`- cleanAppAsync ${projectData.projectName}`)
+  await removeProjectAsync(projectData)
+  const result = await syncGeneratorAsync(projectData.projectName)
+  testDebug(result.stdout)
+  expect(result.stdout).toMatch(/DELETE (firebase)(\S*)(.json)/)
+  expectStrings(result.stdout, [
+    `This workspace has ${options.appsRemaining} firebase apps and ${options.functionsRemaining} firebase functions`,
+    `CHANGE Firebase config '${projectData.configName}' is no longer referenced by any firebase app, deleted`
+  ])
+}  
+
+export async function cleanFunctionAsync(projectData: ProjectData) {
+  testDebug(`- cleanFunctionAsync ${projectData.projectName}`)
+  await removeProjectAsync(projectData)
+}  
+
+
+
+
+/**
+ * Generate test project data
+ * Note: call this function AFTER initial app firebase.json has been created in order to have a
+ *  correct configName
+ * @param name - project name (cannot be camel case)
+ * @param dir - project dir
+ * @returns - asset locations for this project
+ */
+export function getProjectData(
+  type: 'libs' | 'apps',
+  name: string,
+  options?: { dir?: string; customConfig?: boolean },
+): ProjectData {
+  // Nx16.8.0+ no longer adds the 'apps' or 'libs' prefix in the project name
+  // --directory=${projectData.projectDir} is used instead
+  // see https://nx.dev/deprecated/as-provided-vs-derived
+
+  // we want to maintain the kebab-case name for the project/dir
+  // but we need to ensure the 'apps' or 'libs' prefix is added to the --directory
+
+  // const dir = options?.dir ? `${names(options.dir).fileName}` : ''
+  const d = options?.dir ? `${names(options.dir).fileName}` : ''
+  const dir = joinPathFragments(type, d)
+
+  // project name is kebab case dir + name
+  const n = names(name).fileName
+  const prefix = options?.dir ? `${d}-` : ''
+  const projectName = `${prefix}${n}`
+
+  const projectDir = joinPathFragments(dir, n)
+
+  const srcDir = joinPathFragments(projectDir, 'src')
+  const mainTsPath = joinPathFragments(srcDir, 'main.ts')
+  const distDir = joinPathFragments('dist', projectDir)
+
+  return {
+    name, // name passed to generator
+    directory: dir, // --directory option required for generators
+    projectName, // project name
+    projectDir,
+    srcDir,
+    distDir,
+    mainTsPath,
+    npmScope: `${NPM_SCOPE}/${projectName}`,
+    configName: options?.customConfig
+      ? `firebase.${projectName}.json`
+      : 'firebase.json',
+  }
+}
+
+const IMPORT_MATCH = `import * as logger from "firebase-functions/logger";`
+
+export function getMainTs() {
+  return `
+/**
+ * Import function triggers from their respective submodules:
+ *
+ * import {onCall} from "firebase-functions/v2/https";
+ * import {onDocumentWritten} from "firebase-functions/v2/firestore";
+ *
+ * See a full list of supported triggers at https://firebase.google.com/docs/functions
+ */
+
+import {onRequest} from "firebase-functions/v2/https";
+${IMPORT_MATCH}
+
+// Start writing functions
+// https://firebase.google.com/docs/functions/typescript
+
+export const helloWorld = onRequest((request, response) => {
+  logger.info("Hello logs!", {structuredData: true});
+  response.send("Hello from Firebase!");
+});
+`
+}
+
+
+/**
+ * return the import function for a generated library
+ */
+export function getLibImport(projectData: ProjectData) {
+  // convert kebab-case project name to camelCase library import
+  const libName = projectData.projectName.split('-').map((part, index) => index > 0 ? part[0].toUpperCase() + part.substring(1) : part).join('')
+  return libName
+}
+
+export function addImport(mainTs: string, addition: string) {
+  const replaced = mainTs.replace(IMPORT_MATCH, `${IMPORT_MATCH}\n${addition}`)
+  return replaced
+}
+
+
+export function expectedAppProjectTargets(appProject: ProjectData) {
+  return {
+    build: {
+      executor: 'nx:run-commands',
+      options: {
+        command: `echo Build succeeded.`,
+      },
+    },
+    watch: {
+      executor: 'nx:run-commands',
+      options: {
+        command: `nx run-many --targets=build --projects=tag:firebase:dep:${appProject.projectName} --parallel=100 --watch`,
+      },
+    },
+    lint: {
+      executor: 'nx:run-commands',
+      options: {
+        command: `nx run-many --targets=lint --projects=tag:firebase:dep:${appProject.projectName} --parallel=100`,
+      },
+    },
+    test: {
+      executor: 'nx:run-commands',
+      options: {
+        command: `nx run-many --targets=test --projects=tag:firebase:dep:${appProject.projectName} --parallel=100`,
+      },
+    },
+    firebase: {
+      executor: 'nx:run-commands',
+      options: {
+        command: `firebase --config=firebase.json`,
+      },
+      configurations: {
+        production: {
+          command: `firebase --config=firebase.json`,
+        },
+      },
+    },
+    killports: {
+      executor: 'nx:run-commands',
+      options: {
+        command: `kill-port --port 9099,5001,8080,9000,5000,8085,9199,9299,4000,4400,4500`,
+      },
+    },
+    getconfig: {
+      executor: 'nx:run-commands',
+      options: {
+        command: `nx run ${appProject.projectName}:firebase functions:config:get > ${appProject.projectDir}/environment/.runtimeconfig.json`,
+      },
+    },
+    emulate: {
+      executor: 'nx:run-commands',
+      options: {
+        commands: [
+          `nx run ${appProject.projectName}:killports`,
+          `nx run ${appProject.projectName}:firebase emulators:start --import=${appProject.projectDir}/.emulators --export-on-exit`,
+        ],
+        parallel: false,
+      },
+    },
+    serve: {
+      executor: '@simondotm/nx-firebase:serve',
+      options: {
+        commands: [
+          `nx run ${appProject.projectName}:watch`,
+          `nx run ${appProject.projectName}:emulate`,
+        ],
+      },
+    },
+    deploy: {
+      executor: 'nx:run-commands',
+      dependsOn: ['build'],
+      options: {
+        command: `nx run ${appProject.projectName}:firebase deploy`,
+      },
+    },
+  }
+}
+
+
+export function expectedFunctionProjectTargets(
+  functionProject: ProjectData,
+  appProject: ProjectData,
+) {
+  return {
+    build: {
+      executor: '@nx/esbuild:esbuild',
+      outputs: ['{options.outputPath}'],
+      options: {
+        platform: 'node',
+        outputPath: `dist/${functionProject.projectDir}`,
+        main: `${functionProject.projectDir}/src/main.ts`,
+        tsConfig: `${functionProject.projectDir}/tsconfig.app.json`,
+        assets: [
+          `${functionProject.projectDir}/src/assets`,
+          {
+            glob: '**/*',
+            input: `${appProject.projectDir}/environment`,
+            output: '.',
+          },
+        ],
+        generatePackageJson: true,
+        bundle: true,
+        dependenciesFieldType: 'dependencies',
+        format: ['esm'],
+        thirdParty: false,
+        target: 'node16',
+        esbuildOptions: {
+          logLevel: 'info',
+        },
+      },
+    },
+    deploy: {
+      executor: 'nx:run-commands',
+      options: {
+        command: `nx run ${appProject.projectName}:deploy --only functions:${functionProject.projectName}`,
+      },
+      dependsOn: ['build'],
+    },
+    lint: {
+      executor: '@nx/linter:eslint',
+      outputs: ['{options.outputFile}'],
+      options: {
+        lintFilePatterns: [`${functionProject.projectDir}/**/*.ts`],
+      },
+    },
+    test: {
+      executor: '@nx/jest:jest',
+      outputs: [`{workspaceRoot}/coverage/{projectRoot}`],
+      options: {
+        jestConfig: `${functionProject.projectDir}/jest.config.ts`,
+        passWithNoTests: true,
+      },
+      configurations: {
+        ci: {
+          ci: true,
+          codeCoverage: true,
+        },
+      },      
+    },
+  }
+}
+
+
+export function validateProjectConfig(appProject: ProjectData) {
+    const project = readJson(
+      `${appProject.projectDir}/project.json`,
+    )
+    // expect(project.root).toEqual(`apps/${projectName}`)
+    expect(project.targets).toEqual(
+      expect.objectContaining(expectedAppProjectTargets(appProject)),
+    )
+}
+
+
+export function validateFunctionConfig(functionProject: ProjectData, appProject: ProjectData) {
+    const project = readJson(
+      `${functionProject.projectDir}/project.json`,
+    )
+    // expect(project.root).toEqual(`apps/${projectName}`)
+    expect(project.targets).toEqual(
+      expect.objectContaining(expectedFunctionProjectTargets(functionProject, appProject)),
+    )
+}
+
+
+
